+++ flask_template/util/config/smorest_config.py
@@ -2,11 +2,19 @@ class SmorestProductionConfig:
     OPENAPI_VERSION = "3.0.2"
     OPENAPI_JSON_PATH = "api-spec.json"
     OPENAPI_URL_PREFIX = "/api"
-    OPENAPI_REDOC_PATH = "/redoc"
+
+    # OpenAPI Documentation renderers:
+    OPENAPI_REDOC_PATH = "/redoc/"
     OPENAPI_REDOC_URL = (
         "https://cdn.jsdelivr.net/npm/redoc@next/bundles/redoc.standalone.js"
     )
-    OPENAPI_SWAGGER_UI_PATH = "/swagger-ui"
+
+    OPENAPI_RAPIDOC_PATH = "/rapidoc/"
+    OPENAPI_RAPIDOC_URL = "https://cdn.jsdelivr.net/npm/rapidoc/dist/rapidoc-min.js"
+    # mor config options: https://mrin9.github.io/RapiDoc/api.html
+    OPENAPI_RAPIDOC_CONFIG = {"use-path-in-nav-bar": "true"}
+    
+    OPENAPI_SWAGGER_UI_PATH = "/swagger-ui/"
     OPENAPI_SWAGGER_UI_URL = "https://cdn.jsdelivr.net/npm/swagger-ui-dist/"