+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Update flask-smorest to version 0.31.1 to enable rapidoc OpenAPI documentation renderer
+
 
 ## [v0.2.0]